pyproject.toml
[tool.poetry]
name = "flaskstartertemplate"
version = "0.1.0"
description = ""
authors = ["Ajawu David <ajawudavid@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
Flask = "^2.0.1"
python-decouple = "^3.4"
celery = "^5.1.2"
Flask-Migrate = "^3.1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
